requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
#apache-airflow-providers-google==10.3.0
#google-cloud-bigquery-storage
#soda-core-bigquery==3.0.45
#astro-run-dag # This package is needed for the astro run command. It will be removed before a deploy

# REMOVE apache-airflow-providers-google==10.3.0
# REMOVE soda-core-bigquery==3.0.45
astronomer-cosmos[dbt-bigquery]==1.0.3 # install google + cosmos + dbt
protobuf==3.20.0